2023/6/10 下午4:13

+++ src/main/java/com/Li/esp32mqttserver/domain/Good.java
@@ -21,5 +21,5 @@ public class Good {
     private String name;
     private Double price;
     private String goodsDetails;
-    private Integer stock;
+    private Long stock;
 }
